feat(build): implement build scripts for multiple packages and restructure build command. Closes #87
